(#9) Добавлен справочник тип говорящих. Доступно 3 типа: С ролями, Без ролей, Все поля. Для каждого проекта выбирается типа говорящих, который определяет используемые поля в таблице говорящих.

+++ qstatementwindow.h
@@ -59,7 +59,7 @@ private slots:
   void monitorCheck(MonitorState &state, QString &description);
 
 public slots:
-  void onProjectLoaded(qlonglong oldId, qlonglong newId);
+  void onProjectLoaded(int oldId, int newId);
 private:
   static QStatementWindow* singletonWindow;
   QDataModule* dm;
